Decided to go with counting instead of radix. initial version of counting working.

+++ project4/Merge.cpp
@@ -1,6 +1,6 @@
 #include "Merge.h"
 
-Merge::Merge(int toSort[], int count) {
+Merge::Merge(int toSort[], long count) {
     this->count = count; // set count
     data = new int[count]; // create array of size count
     for (int i = 0; i < count; i++) {
@@ -13,17 +13,17 @@ Merge::Merge(int toSort[], int count) {
     swapCount = 0;
 }
 
-void Merge::merge(int first, int mid, int last) {
+void Merge::merge(long first, long mid, long last) {
     int tempArray[count]; // Temporary array
                                 // Initialize the local indices to indicate the subarrays
-    int first1 = first;           // Beginning of first subarray
-    int last1 = mid;              // End of first subarray
-    int first2 = mid + 1;         // Beginning of second subarray
-    int last2 = last;             // End of second subarray
+    long first1 = first;           // Beginning of first subarray
+    long last1 = mid;              // End of first subarray
+    long first2 = mid + 1;         // Beginning of second subarray
+    long last2 = last;             // End of second subarray
 
     // While both subarrays are not empty, copy the
     // smaller item into the temporary array
-    int index = first1;           // Next available location in tempArray
+    long index = first1;           // Next available location in tempArray
     while ((first1 <= last1) && (first2 <= last2)) {
         // At this point, tempArray[first..index–1] is in order
         if (data[first1] <= data[first2]) {
@@ -60,7 +60,9 @@ void Merge::merge(int first, int mid, int last) {
 
 }  // end merge
 
-void Merge::sort(int first, int last) {
+void Merge::sort(long first, long last) {
+
+    // cout << "first: " << first << ", last: " << last << endl;
     /* Reset comparison and swap count in case */
     comparisonCount = 0;
     swapCount = 0;
@@ -69,7 +71,7 @@ void Merge::sort(int first, int last) {
 
     if (first < last) {
         // Sort each half
-        int mid = first + (last - first) / 2; // Index of midpoint
+        long mid = first + (last - first) / 2; // Index of midpoint
 
         // Sort left half theArray[first..mid]
         sort(first, mid);
